Use bare json

+++ app/controllers/api/sessions_controller.rb
@@ -10,24 +10,20 @@ class Api::SessionsController < Api::ApplicationController
     if user
       access_token = AccessToken.create(user: user)
       render json: {
-        logged_in: true,
         token: access_token.token,
         user: UserSerializer.new(user)
-      }, status: 201, root: false
+      },
+      status: 201,
+      root: false
     else
-      render json: { logged_in: false }, status: 422
+      head 422
     end
   end
 
   def destroy
     token = request.headers["Access-Token"]
     access_token = AccessToken.find_token(token)
-
-    if access_token
-      access_token.destroy
-      render json: { logged_out: true }, status: 200
-    else
-      render json: { logged_out: false }, status: 200
-    end
+    access_token.destroy if access_token
+    head 200
   end
 end
